Make pipong example self-contained.
(The code could be further cleaned up, but this is just merging the two
files together and fixing the docs/imports.)
